Added voice trigger ("show me a demo"); updated code style to match that of other samples.
