+++ AndroidManifest.xml
@@ -1,6 +1,21 @@
 <?xml version="1.0" encoding="utf-8"?>
+<!-- Copyright (C) 2013 Google Inc.
+
+     Licensed under the Apache License, Version 2.0 (the "License");
+     you may not use this file except in compliance with the License.
+     You may obtain a copy of the License at
+
+          http://www.apache.org/licenses/LICENSE-2.0
+
+     Unless required by applicable law or agreed to in writing, software
+     distributed under the License is distributed on an "AS IS" BASIS,
+     WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+     See the License for the specific language governing permissions and
+     limitations under the License.
+-->
+
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.google.glass.samples.waveform"
+    package="com.google.android.glass.sample.waveform"
     android:versionCode="1"
     android:versionName="1.0" >
 
@@ -12,15 +27,17 @@
 
     <application
         android:allowBackup="true"
-        android:label="@string/app_name"
-        android:theme="@android:style/Theme.NoTitleBar.Fullscreen" >
+        android:label="@string/app_name" >
+
         <activity
-            android:name="com.google.glass.samples.waveform.WaveformActivity"
-            android:label="@string/app_name" >
+            android:name="com.google.android.glass.sample.waveform.WaveformActivity" >
             <intent-filter>
-                <action android:name="android.intent.action.MAIN" />
+                <action android:name="com.google.android.glass.action.VOICE_TRIGGER" />
             </intent-filter>
+            <meta-data android:name="com.google.android.glass.VoiceTrigger"
+                android:resource="@xml/trigger_show_me_a_demo" />
         </activity>
+
     </application>
 
 </manifest>
