+++ res/layout/layout_waveform.xml
@@ -1,23 +1,36 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!-- Copyright (C) 2013 Google Inc.
+
+     Licensed under the Apache License, Version 2.0 (the "License");
+     you may not use this file except in compliance with the License.
+     You may obtain a copy of the License at
+
+          http://www.apache.org/licenses/LICENSE-2.0
+
+     Unless required by applicable law or agreed to in writing, software
+     distributed under the License is distributed on an "AS IS" BASIS,
+     WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+     See the License for the specific language governing permissions and
+     limitations under the License.
+-->
+
 <FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:keepScreenOn="true"
-    android:background="@color/black"
-    tools:context=".WaveformActivity" >
+    android:layout_height="match_parent" >
 
-    <com.google.glass.samples.waveform.WaveformView
-        android:id="@+id/waveformView"
+    <com.google.android.glass.sample.waveform.WaveformView
+        android:id="@+id/waveform_view"
         android:layout_width="match_parent"
         android:layout_height="match_parent" />
 
     <TextView
-        android:id="@+id/decibelView"
+        android:id="@+id/decibel_view"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_gravity="bottom|right"
         android:layout_marginRight="@dimen/footer_right_margin"
         android:layout_marginBottom="@dimen/footer_bottom_margin"
-        android:textSize="@dimen/footer_text_size" />
-    
+        android:textAppearance="?android:attr/textAppearanceSmall" />
+
 </FrameLayout>
