+++ res/values/dimens.xml
@@ -1,6 +1,22 @@
 <?xml version="1.0" encoding="utf-8"?>
+<!-- Copyright (C) 2013 Google Inc.
+
+     Licensed under the Apache License, Version 2.0 (the "License");
+     you may not use this file except in compliance with the License.
+     You may obtain a copy of the License at
+
+          http://www.apache.org/licenses/LICENSE-2.0
+
+     Unless required by applicable law or agreed to in writing, software
+     distributed under the License is distributed on an "AS IS" BASIS,
+     WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+     See the License for the specific language governing permissions and
+     limitations under the License.
+-->
+
 <resources>
+
     <dimen name="footer_right_margin">40px</dimen>
     <dimen name="footer_bottom_margin">29px</dimen>
-    <dimen name="footer_text_size">30px</dimen>
+
 </resources>
