+++ res/xml/trigger_show_me_a_demo.xml
@@ -14,11 +14,4 @@
      limitations under the License.
 -->
 
-<resources>
-
-    <string name="app_name">Waveform</string>
-    <string name="show_me_a_demo">show me a demo</string>
-
-    <string name="decibel_format">%1$.1f dB</string>
-
-</resources>
+<trigger keyword="@string/show_me_a_demo" />
